requirements.txt
alembic==1.0.1
autopep8==1.3.5
certifi==2018.10.15
Click==7.0
cycler==0.10.0
dlib==19.15.0
flake8==3.5.0
Flask==1.0.2
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
imutils==0.4.6
Jinja2==2.10
kiwisolver==1.0.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mkl-random==1.0.1
numpy==1.15.2
Pillow==5.3.0
psycopg2==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
pyparsing==2.2.0
pyqtgraph==0.10.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
SQLAlchemy==1.2.12
tornado==5.1
webcolors==1.8.1
Werkzeug==0.14.1
